Move the setting-related component into the settings namespace.
